+++ aldjemy/orm.py
@@ -1,11 +1,12 @@
 from sqlalchemy import orm
+from django.apps import apps
 from django.db.models.fields.related import ForeignKey, OneToOneField, ManyToManyField
 from django.db import connections, router
 from django.db.backends import signals
 from django.conf import settings
 
 from .core import get_engine
-from .table import get_all_django_models, generate_tables
+from .table import generate_tables
 
 
 def get_session(alias="default", recreate=False, **kwargs):
@@ -109,7 +110,11 @@ def construct_models(metadata):
     if not metadata.tables:
         generate_tables(metadata)
     tables = metadata.tables
-    models = [model for model in get_all_django_models() if not model._meta.proxy]
+    models = [
+        model
+        for model in apps.get_models(include_auto_created=True)
+        if not model._meta.proxy
+    ]
 
     sa_models = {}